+++ tests/fixtures/golden-install-parity/antigravity.json
@@ -289,7 +289,7 @@
   "gsd-core/workflows/spike.md": "0f9a81bcf4573195",
   "gsd-core/workflows/stats.md": "a20eb078d2ab11be",
   "gsd-core/workflows/sync-skills.md": "8326a7ff0411b077",
-  "gsd-core/workflows/thread.md": "03a527a71b8fab12",
+  "gsd-core/workflows/thread.md": "3fb6b552e45fedbd",
   "gsd-core/workflows/transition.md": "a7a5fe4040084308",
   "gsd-core/workflows/ui-phase.md": "652785fbba26e80c",
   "gsd-core/workflows/ui-review.md": "816b2bde136157f9",
