+++ tests/fixtures/golden-install-parity/claude.json
@@ -288,7 +288,7 @@
   "gsd-core/workflows/spike.md": "aae8bcad15642645",
   "gsd-core/workflows/stats.md": "17b4f2059f4b4ef2",
   "gsd-core/workflows/sync-skills.md": "b505e6f8331c0918",
-  "gsd-core/workflows/thread.md": "75df5cc71f72c33d",
+  "gsd-core/workflows/thread.md": "5ad0e0d5ce7e0d11",
   "gsd-core/workflows/transition.md": "96ce39403ca69594",
   "gsd-core/workflows/ui-phase.md": "790e5982e5b715c3",
   "gsd-core/workflows/ui-review.md": "51945fda8e931f99",
