+++ tests/fixtures/golden-install-parity/cline.json
@@ -292,7 +292,7 @@
   "gsd-core/workflows/spike.md": "204e742c846ee0d9",
   "gsd-core/workflows/stats.md": "5cfea82b894eee3c",
   "gsd-core/workflows/sync-skills.md": "b505e6f8331c0918",
-  "gsd-core/workflows/thread.md": "5ae4c3141bdedd88",
+  "gsd-core/workflows/thread.md": "4a009fd2cc4387a7",
   "gsd-core/workflows/transition.md": "fe82e77df8dceb1b",
   "gsd-core/workflows/ui-phase.md": "06f1f80de620a319",
   "gsd-core/workflows/ui-review.md": "0af83311f5e41f48",
