+++ tests/fixtures/golden-install-parity/codebuddy.json
@@ -358,7 +358,7 @@
   "gsd-core/workflows/spike.md": "0051a7e2193a7522",
   "gsd-core/workflows/stats.md": "01c24349370a0e6d",
   "gsd-core/workflows/sync-skills.md": "b505e6f8331c0918",
-  "gsd-core/workflows/thread.md": "c26ca43fdf928d46",
+  "gsd-core/workflows/thread.md": "6d075b5d26500e9d",
   "gsd-core/workflows/transition.md": "eee3435817fab185",
   "gsd-core/workflows/ui-phase.md": "93ea0d0dfcb2a1e8",
   "gsd-core/workflows/ui-review.md": "1ad3654435000881",
