+++ tests/fixtures/golden-install-parity/codex.json
@@ -324,7 +324,7 @@
   "gsd-core/workflows/spike.md": "c2f115f0d3251654",
   "gsd-core/workflows/stats.md": "0d7449acf349feec",
   "gsd-core/workflows/sync-skills.md": "e2b793963799f8ce",
-  "gsd-core/workflows/thread.md": "d3f768ce0f4b4a4d",
+  "gsd-core/workflows/thread.md": "ee872031abd592e0",
   "gsd-core/workflows/transition.md": "c4bded570fafe712",
   "gsd-core/workflows/ui-phase.md": "b373c964b222324c",
   "gsd-core/workflows/ui-review.md": "a9b2cbba80482cd8",
