+++ tests/fixtures/golden-install-parity/copilot.json
@@ -290,7 +290,7 @@
   "gsd-core/workflows/spike.md": "716d74cdb2e39a3e",
   "gsd-core/workflows/stats.md": "49085df6d4793df3",
   "gsd-core/workflows/sync-skills.md": "eca50ffe8320dba8",
-  "gsd-core/workflows/thread.md": "4c44f10d41740f1d",
+  "gsd-core/workflows/thread.md": "d53698a91bbbe33c",
   "gsd-core/workflows/transition.md": "724b6e9b34d85f26",
   "gsd-core/workflows/ui-phase.md": "9fefa0db49f2aa3f",
   "gsd-core/workflows/ui-review.md": "731bca05f9770a86",
