+++ tests/fixtures/golden-install-parity/cursor.json
@@ -358,7 +358,7 @@
   "gsd-core/workflows/spike.md": "9e37f8067adf87b7",
   "gsd-core/workflows/stats.md": "17b4f2059f4b4ef2",
   "gsd-core/workflows/sync-skills.md": "b505e6f8331c0918",
-  "gsd-core/workflows/thread.md": "06e31fdaf02ccabc",
+  "gsd-core/workflows/thread.md": "cfc42471f9001ba8",
   "gsd-core/workflows/transition.md": "96ce39403ca69594",
   "gsd-core/workflows/ui-phase.md": "dea2f2d43a43e0d0",
   "gsd-core/workflows/ui-review.md": "6b16a7f7783be471",
