+++ tests/fixtures/golden-install-parity/gemini.json
@@ -358,7 +358,7 @@
   "gsd-core/workflows/spike.md": "c9482514e665bcac",
   "gsd-core/workflows/stats.md": "01c24349370a0e6d",
   "gsd-core/workflows/sync-skills.md": "b505e6f8331c0918",
-  "gsd-core/workflows/thread.md": "c26ca43fdf928d46",
+  "gsd-core/workflows/thread.md": "6d075b5d26500e9d",
   "gsd-core/workflows/transition.md": "69143573741c52e4",
   "gsd-core/workflows/ui-phase.md": "bb5167948032872e",
   "gsd-core/workflows/ui-review.md": "d256bec482e67af8",
