+++ tests/fixtures/golden-install-parity/hermes.json
@@ -289,7 +289,7 @@
   "gsd-core/workflows/spike.md": "80844a3c05339fdb",
   "gsd-core/workflows/stats.md": "01289f444b66913d",
   "gsd-core/workflows/sync-skills.md": "b505e6f8331c0918",
-  "gsd-core/workflows/thread.md": "7af5046e18c81ee9",
+  "gsd-core/workflows/thread.md": "596bf37dcea2ce1e",
   "gsd-core/workflows/transition.md": "adab3f53afd5d8aa",
   "gsd-core/workflows/ui-phase.md": "dd213d91b5c258a5",
   "gsd-core/workflows/ui-review.md": "83d6d3b1f26597ff",
