+++ tests/fixtures/golden-install-parity/kilo.json
@@ -358,7 +358,7 @@
   "gsd-core/workflows/spike.md": "a782dd863771fe0a",
   "gsd-core/workflows/stats.md": "17b4f2059f4b4ef2",
   "gsd-core/workflows/sync-skills.md": "dc7b8b015afa4b3b",
-  "gsd-core/workflows/thread.md": "75df5cc71f72c33d",
+  "gsd-core/workflows/thread.md": "5ad0e0d5ce7e0d11",
   "gsd-core/workflows/transition.md": "9040a76741f12de6",
   "gsd-core/workflows/ui-phase.md": "33f7113e91c2bfe0",
   "gsd-core/workflows/ui-review.md": "013272816a291305",
