+++ tests/fixtures/golden-install-parity/opencode.json
@@ -358,7 +358,7 @@
   "gsd-core/workflows/spike.md": "df52b9f31a72d204",
   "gsd-core/workflows/stats.md": "598b1bb510ed0451",
   "gsd-core/workflows/sync-skills.md": "7e2c138cdbef4282",
-  "gsd-core/workflows/thread.md": "c11265f0fa0a95d3",
+  "gsd-core/workflows/thread.md": "f4f8312f744c9cba",
   "gsd-core/workflows/transition.md": "2d6739f730c3ea10",
   "gsd-core/workflows/ui-phase.md": "38dac814d2d03d6f",
   "gsd-core/workflows/ui-review.md": "3d972d3bd30527b3",
