+++ tests/fixtures/golden-install-parity/qwen.json
@@ -289,7 +289,7 @@
   "gsd-core/workflows/spike.md": "52fcd95f7b343232",
   "gsd-core/workflows/stats.md": "7ffa072290ebcae3",
   "gsd-core/workflows/sync-skills.md": "b505e6f8331c0918",
-  "gsd-core/workflows/thread.md": "d85a53d03a3167e0",
+  "gsd-core/workflows/thread.md": "c10177767648ea7f",
   "gsd-core/workflows/transition.md": "1f0a1478d75d89ec",
   "gsd-core/workflows/ui-phase.md": "1c91323daf451053",
   "gsd-core/workflows/ui-review.md": "f1f82d8257e910cc",
