+++ tests/fixtures/golden-install-parity/trae.json
@@ -289,7 +289,7 @@
   "gsd-core/workflows/spike.md": "48df8b626cbd378d",
   "gsd-core/workflows/stats.md": "18089135bc41f1b4",
   "gsd-core/workflows/sync-skills.md": "b505e6f8331c0918",
-  "gsd-core/workflows/thread.md": "9fe3ec12e491bec3",
+  "gsd-core/workflows/thread.md": "61e0eee731434ea9",
   "gsd-core/workflows/transition.md": "b49ddd9247f804e7",
   "gsd-core/workflows/ui-phase.md": "a725ebdd4f47fb97",
   "gsd-core/workflows/ui-review.md": "b3221cac976daffa",
