+++ tests/fixtures/golden-install-parity/windsurf.json
@@ -289,7 +289,7 @@
   "gsd-core/workflows/spike.md": "aa5934044317ba7a",
   "gsd-core/workflows/stats.md": "c49d3de15375d04b",
   "gsd-core/workflows/sync-skills.md": "b505e6f8331c0918",
-  "gsd-core/workflows/thread.md": "5ad6ddb308909770",
+  "gsd-core/workflows/thread.md": "3e9fcfbb254a31a7",
   "gsd-core/workflows/transition.md": "9a49f9c48805f666",
   "gsd-core/workflows/ui-phase.md": "54e01f1973450a3c",
   "gsd-core/workflows/ui-review.md": "fea93c281767f8d7",
