+++ tests/workflow-size-baseline.json
@@ -78,7 +78,7 @@
   "spike.md": 24517,
   "stats.md": 6718,
   "sync-skills.md": 6125,
-  "thread.md": 12400,
+  "thread.md": 12464,
   "transition.md": 22016,
   "ui-phase.md": 15477,
   "ui-review.md": 11172,
